+++ lib/cripto_details/detail_screen.dart
