+++ lib/main.dart
@@ -1,5 +1,4 @@
-import 'screens/wallet/wallet_screen.dart';
-
+import 'wallet/view/wallet_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -17,9 +16,9 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.pink,
       ),
-      home: WalletScreen(),
+      home: const WalletScreen(),
     );
   }
 }
