+++ lib/shared/widgets/default_title.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-class TitleWallet extends StatelessWidget {
+class DefaultTitle extends StatelessWidget {
   final String title;
 
-  const TitleWallet(this.title, {Key? key}) : super(key: key);
+  const DefaultTitle(this.title, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
