+++ lib/shared/widgets/navbar_wallet.dart
