+++ lib/shared/widgets/number_formatter.dart
@@ -1,0 +1,5 @@
+import 'package:intl/intl.dart';
+
+final formatReais = NumberFormat('#,###.00#', 'pt_BR');
+final formatCriptoAbrev = NumberFormat('#.##', 'en_US');
+final formatCriptoCompleto = NumberFormat('#0.#########', 'en_US');
