+++ lib/wallet/view/wallet_screen.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+import '../../shared/widgets/navbar_wallet.dart';
+import '../widgets/wallet_body.dart';
+
+class WalletScreen extends StatefulWidget {
+  const WalletScreen({Key? key}) : super(key: key);
+
+  @override
+  State<WalletScreen> createState() => _WalletScreenState();
+}
+
+class _WalletScreenState extends State<WalletScreen> {
+  late int selectedIndex = 0;
+
+  void onItemTapped(int index) {
+    setState(
+      () {
+        selectedIndex = index;
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: WalletBody(),
+      bottomNavigationBar: NavbarWallet(
+        selectedIndex: selectedIndex,
+        onItemTapped: onItemTapped,
+      ),
+    );
+  }
+}
