+++ lib/wallet/widgets/total_value.dart
@@ -1,8 +1,9 @@
-import 'package:crypto_listing/shared/template/wallet_providers.dart';
+import '../../shared/providers/wallet_providers.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:intl/intl.dart';
+
+import '../../shared/widgets/number_formatter.dart';
 
 class TotalValue extends StatefulHookConsumerWidget {
   final double totalReais;
@@ -17,17 +18,13 @@ class TotalValue extends StatefulHookConsumerWidget {
 }
 
 class _TotalValueState extends ConsumerState<TotalValue> {
-  var format = NumberFormat('#,###.00#', 'pt_BR');
-  // var formatter = NumberFormat.decimalPattern('pt-BR');
-
   @override
   Widget build(BuildContext context) {
     final bool hideWallet = ref.watch(hideWalletStateProvider);
     return Visibility(
       visible: hideWallet,
       replacement: Text(
-        'R\$ ${format.format(widget.totalReais)}',
-        // 'R\$ ${formatter.format(DecimalIntl(Decimal.parse('${widget.totalReais}')))}',
+        'R\$ ${formatReais.format(widget.totalReais)}',
         style:
             GoogleFonts.montserrat(fontSize: 32, fontWeight: FontWeight.w800),
       ),
